ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "AI vision pipeline for water level detection"
requires-python = ">=3.12"
dependencies = [
    "numpy<2",
    "opencv-python<4.13",
    "pillow>=12.2.0",
    "segment-anything",
    "torch>=2.1,<2.3",
    "torchvision>=0.16,<0.18",
    "ultralytics>=8.4.41",
    "hydra-core>=1.3",
    "tqdm",
]

[tool.uv.sources]
segment-anything = { git = "https://github.com/facebookresearch/segment-anything.git" }
